Flipped sign: sensible heat flux and evaporation
I also added a 'standard_name' attribute for these fields so it is obvious to
anyone looking at them that they are positive-up. We probably should add
standard_name attributes to the rest of the fields marking them as positive
down, but since that's the standard most MOM6 fields expect that doesn't seem
as important for now.